Fixed frozen tails and timed pattern release.
It's possible that a light will be left with intensity but have no
remaining patterns -- had to account for this and make sure those lights
always get set to an intensity of 0. Also fixed the problem with the
timed release of patterns (Clock.getElapsedTime() returns the number of
seconds since its last call...).