One geometry for casings (worse than many cubes).
Tried reusing a single geometry for each light's casing (i.e., back and
sides). Each light was thus made up of a mesh/geometry for the casing
and planar mesh/geometry for the emitter face. This performs worse than
if each light is its own CubeGeometry and we update the downward
pointing face on each.